+++ tests/utils.rs
@@ -1,11 +1,27 @@
 use assert_cmd::cargo::cargo_bin;
 use proc_mounts::MountIter;
 use std::{
+    fs,
     path::{Path, PathBuf},
     process, thread,
     time::Duration,
 };
 use tempfile::TempDir;
+use zstd::decode_all;
+
+pub fn sync_file<P>(path: P)
+where
+    P: AsRef<Path>,
+{
+    fs::File::open(path).unwrap().sync_all().unwrap();
+}
+
+pub fn get_compressed_content<P>(path: P) -> String
+where
+    P: AsRef<Path>,
+{
+    String::from_utf8(decode_all(fs::File::open(path).unwrap()).unwrap()).unwrap()
+}
 
 pub struct FuseZstdProcess {
     process: process::Child,
@@ -14,12 +30,14 @@ pub struct FuseZstdProcess {
 }
 
 impl FuseZstdProcess {
-    pub fn new() -> Self {
+    pub fn new(convert: bool) -> Self {
         let data_dir = TempDir::new_in("/tmp/").unwrap();
         let mount_point = TempDir::new_in("/tmp/").unwrap();
         let process = process::Command::new(cargo_bin("fuse-zstd"))
             .args(["--data-dir", data_dir.path().to_str().unwrap()])
             .args(["--mount-point", mount_point.path().to_str().unwrap()])
+            .args(if convert { vec!["--convert"] } else { vec![] })
+            .arg("-vvv")
             .spawn()
             .unwrap();